// src/types/api/bindings/CreateUserSessionResponse.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { PartialUser } from './PartialUser';

/**
 * Defines a CreateUserSessionResponse structure which is used to return session information
 * after creation of a session
 *
 * May contain partial user information if the session was created via OAuth2 login
 */
export type CreateUserSessionResponse = {
	/**
	 * The ID of the user who created the session
	 */
	user_id: string;
	/**
	 * The token of the session
	 */
	token: string;
	/**
	 * The ID of the session
	 */
	session_id: string;
	/**
	 * The time the session expires
	 */
	expiry: string;
	/**
	 * The user who created the session (only sent on OAuth2 login)
	 */
	user: PartialUser | null;
};